Fix w# argument.
git-svn-id: http://svn.python.org/projects/python/branches/ssize_t@41874 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Objects/fileobject.c
@@ -854,7 +854,7 @@ static PyObject *
 file_readinto(PyFileObject *f, PyObject *args)
 {
 	char *ptr;
-	int ntodo;
+	Py_ssize_t ntodo;
 	Py_ssize_t ndone, nnow;
 
 	if (f->f_fp == NULL)
